# swagger.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "Users API",
    "description": "Users API with authentication functionality"
  },
  "host": "api-gfpedro.cloud.okteto.net",
  "tags": [
    {
      "name": "Auth",
      "description": "API for authentication actions and session managment"
    },
    {
      "name": "Users",
      "description": "API for user actions"
    }
  ],
  "schemes": [
    "https"
  ],
  "securityDefinitions": {
    "bearerAuth": {
      "name": "Authorization",
      "type": "apiKey",
      "scheme": "Bearer",
      "in": "header",
      "description": "JWT Authorization header using the Bearer scheme. \r\n\r\n Enter 'Bearer' [space] and then your token in the text input below.\r\n\r\nExample: \"Bearer superSecureAccessToken\""
    }
  },
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/login": {
      "post": {
        "tags": [
          "Auth"
        ],
        "parameters": [
          {
            "name": "userCredentials",
            "in": "body",
            "description": "User's credentials",
            "required": true,
            "schema": {
              "$ref": "#/definitions/UserCredentials"
            }
          }
        ],
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "Session tokens",
            "schema": {
              "$ref": "#/definitions/AuthResponse"
            }
          }
        }
      }
    },
    "/refreshToken": {
      "post": {
        "tags": [
          "Auth"
        ],
        "parameters": [
          {
            "name": "refreshToken",
            "in": "body",
            "description": "Session refresh token",
            "required": true,
            "schema": {
              "$ref": "#/definitions/TokenRefresh"
            }
          }
        ],
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "New access token",
            "schema": {
              "$ref": "#/definitions/TokenRefreshResponse"
            }
          }
        }
      }
    },
    "/isAuthenticated": {
      "post": {
        "tags": [
          "Auth"
        ],
        "parameters": [
          {
            "name": "accessToken",
            "in": "body",
            "description": "Session access token",
            "required": true,
            "schema": {
              "$ref": "#/definitions/TokenCheck"
            }
          }
        ],
        "produces": [
          "application/json"
        ],
        "responses": {
          "204": {
            "description": "Access token is valid"
          },
          "403": {
            "description": "Access token expired or is not valid"
          }
        }
      }
    },
    "/logout": {
      "security": [
        {
          "bearerAuth": []
        }
      ],
      "delete": {
        "tags": [
          "Auth"
        ],
        "parameters": [
          {
            "name": "refreshToken",
            "in": "body",
            "description": "Session refresh token",
            "required": true,
            "schema": {
              "$ref": "#/definitions/TokenCheck"
            }
          }
        ],
        "produces": [
          "application/json"
        ],
        "responses": {
          "204": {
            "description": "Logged out successfully"
          }
        }
      }
    },
    "/api/v1/users": {
      "post": {
        "tags": [
          "Users"
        ],
        "parameters": [
          {
            "name": "user",
            "in": "body",
            "description": "User we want to create",
            "required": true,
            "schema": {
              "$ref": "#/definitions/UserCreate"
            }
          }
        ],
        "produces": [
          "application/json"
        ],
        "responses": {
          "201": {
            "description": "New user is created",
            "schema": {
              "$ref": "#/definitions/UserResponse"
            }
          }
        }
      },
      "get": {
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "tags": [
          "Users"
        ],
        "parameters": [],
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "List of users",
            "schema": {
              "$ref": "#/definitions/UserList"
            }
          }
        }
      }
    },
    "/api/v1/users/{id}": {
      "put": {
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "tags": [
          "Users"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "User identifier",
            "required": true
          },
          {
            "name": "user",
            "in": "body",
            "description": "User information we want to update",
            "schema": {
              "$ref": "#/definitions/UserCreate"
            }
          }
        ],
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "Updated user info",
            "schema": {
              "$ref": "#/definitions/UserResponse"
            }
          }
        }
      },
      "delete": {
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "tags": [
          "Users"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "User identifier",
            "required": true
          }
        ],
        "produces": [
          "application/json"
        ],
        "responses": {
          "204": {
            "description": "User deleted successfully"
          }
        }
      }
    }
  },
  "definitions": {
    "UserCredentials": {
      "required": [
        "username",
        "password"
      ],
      "properties": {
        "username": {
          "type": "string"
        },
        "password": {
          "type": "string"
        }
      }
    },
    "UserCreate": {
      "required": [
        "name",
        "email",
        "password"
      ],
      "properties": {
        "name": {
          "type": "string"
        },
        "email": {
          "type": "string"
        },
        "password": {
          "type": "string"
        }
      }
    },
    "UserResponse": {
      "required": [
        "id",
        "name",
        "email"
      ],
      "properties": {
        "id": {
          "type": "string"
        },
        "name": {
          "type": "string"
        },
        "email": {
          "type": "string"
        }
      }
    },
    "UserList": {
      "type": "array",
      "$ref": "#/definitions/UserResponse"
    },
    "AuthResponse": {
      "properties": {
        "accessToken": {
          "type": "string"
        },
        "refreshToken": {
          "type": "string"
        },
        "userInfo": {
          "$ref": "#/definitions/UserResponse"
        }
      }
    },
    "TokenRefresh": {
      "properties": {
        "refreshToken": {
          "type": "string"
        }
      }
    },
    "TokenRefreshResponse": {
      "properties": {
        "accessToken": {
          "type": "string"
        }
      }
    },
    "TokenCheck": {
      "properties": {
        "token": {
          "type": "string"
        }
      }
    }
  }
}